KA-228 Rename Utils namespace/folder to StaticWrappers
