+++ NotesBackEnd/src/Notes.Services/Notes.Services.csproj
@@ -43,10 +43,10 @@
     <Compile Include="Notes\AddNoteService.cs" />
     <Compile Include="Notes\GetNoteService.cs" />
     <Compile Include="Notes\UpdateNoteService.cs" />
-    <Compile Include="Utils\DateService.cs" />
+    <Compile Include="StaticWrappers\DateService.cs" />
     <Compile Include="Dependency\ServicesTypesBootstrapper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Utils\GuidService.cs" />
+    <Compile Include="StaticWrappers\GuidService.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Notes.Contracts\Notes.Contracts.csproj">
