+++ NotesBackEnd/src/Notes.Services/StaticWrappers/GuidService.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Notes.Contracts.Services.Utils;
 
-namespace Notes.Services.Utils
+namespace Notes.Services.StaticWrappers
 {
     internal class GuidService : IGuidService
     {
